+++ changelog.md
@@ -1,4 +1,4 @@
-# Upcoming
+# 1.0.2.1
 * Added "Damage reduction: " infos to all damage reduction runes
 * Fixed healthy not stacking on multiple pieces
 * Fixed jump meter not raising the GUI elements when rendering